Use Etherscan fetched code on DiscoUI previews
