+++ packages/l2b/src/implementations/discovery-ui/diffovery/DiffoveryController.ts
@@ -1,21 +1,7 @@
-import { Logger } from '@l2beat/backend-tools'
-import {
-  AllProviders,
-  asStructured,
-  codeIsEOA,
-  findLeadingCommentStart,
-  flattenBytecodeEquivalentStartingFrom,
-  getChainConfigs,
-  getChainFullName,
-  getDiscoveryPaths,
-  type IProvider,
-  SQLiteCache,
-} from '@l2beat/discovery'
-import type { ContractSource } from '@l2beat/discovery/dist/utils/IEtherscanClient'
-import { HttpClient } from '@l2beat/shared'
-import { assert, ChainSpecificAddress, formatJson } from '@l2beat/shared-pure'
+import { findLeadingCommentStart } from '@l2beat/discovery'
+import { assert, type ChainSpecificAddress } from '@l2beat/shared-pure'
 import { type ASTNode, parse } from '@mradomski/fast-solidity-parser'
-import { BigNumber, utils } from 'ethers'
+import type { FlatSourceClient } from './FlatSourceClient'
 
 type DiffoveryResult = {
   name: string
@@ -23,144 +9,33 @@ type DiffoveryResult = {
 }
 
 export class DiffoveryController {
-  private cache = new Map<string, DiffoveryResult>()
-  private httpClient = new HttpClient()
-  private allProviders: AllProviders
+  constructor(private readonly flatSourceClient: FlatSourceClient) {}
 
-  constructor() {
-    const paths = getDiscoveryPaths()
-    const cache = new SQLiteCache(paths.cache)
-
-    this.allProviders = new AllProviders(
-      getChainConfigs(),
-      this.httpClient,
-      cache,
-      Logger.SILENT,
-    )
-  }
-
-  async getClient(shortChain: string): Promise<IProvider> {
-    const chain = getChainFullName(shortChain)
-    const blockNumber = await this.allProviders.getLatestBlockNumber(chain)
-    return this.allProviders.getByBlockNumber(chain, blockNumber)
-  }
-
-  async handle(
-    address: ChainSpecificAddress,
-  ): Promise<DiffoveryResult | undefined> {
-    const chain = ChainSpecificAddress.chain(address)
-    const client = await this.getClient(chain)
-    const cacheKey = address.toString()
-    const cached = this.cache.get(cacheKey)
-    if (cached !== undefined) {
-      return cached
-    }
-
-    const code = await client.getBytecode(address)
-    if (codeIsEOA(code)) {
-      const result = {
-        name: 'EOA',
-        sources: {
-          'L2BEAT-EOA': '// NOTE: Address is an EOA, does not have source code',
-        },
-      }
-      this.cache.set(cacheKey, result)
-      return result
-    }
-
-    const source = await client.getSource(address)
-    const flat = this.handleSource(source)
-    if (source.isVerified) {
-      this.cache.set(cacheKey, {
-        name: source.name,
-        sources: flat,
-      })
-    }
-
-    return {
-      name: source.name.length > 0 ? source.name : 'Unknown',
-      sources: flat,
+  async handle(address: ChainSpecificAddress): Promise<DiffoveryResult> {
+    const result = await this.flatSourceClient.getFlat(address)
+    switch (result.kind) {
+      case 'eoa':
+        return { name: result.name, sources: { 'L2BEAT-EOA': result.flat } }
+      case 'unverified':
+        return {
+          name: result.name,
+          sources: { 'L2BEAT-UNVERIFIED': result.flat },
+        }
+      case 'non-solidity':
+        return {
+          name: result.name,
+          sources: { 'L2BEAT-NON-SOLIDITY': result.flat },
+        }
+      case 'flat':
+        return {
+          name: result.name,
+          sources: {
+            '#ConstructorArguments': result.constructorArguments,
+            ...splitFlatSolidity(result.flat),
+          },
+        }
     }
   }
-
-  private handleSource(source: ContractSource): Record<string, string> {
-    if (!source.isVerified) {
-      return {
-        'L2BEAT-UNVERIFIED':
-          '// NOTE: Source code for this address has not been verified',
-      }
-    }
-    const input = Object.entries(source.files)
-      .map(([fileName, content]) => ({
-        path: fileName,
-        content,
-      }))
-      .filter((e) => e.path.endsWith('.sol'))
-
-    if (input.length === 0) {
-      return {
-        'L2BEAT-NON-SOLIDITY':
-          '// NOTE: DIFFOVERY currently does not support non-solidity contracts',
-      }
-    }
-
-    const flat = flattenBytecodeEquivalentStartingFrom(
-      source.name,
-      source.rootFile,
-      input,
-      source.remappings,
-      { includeAll: true },
-    )
-    const parts = splitFlatSolidity(flat)
-
-    const result = {
-      ['#ConstructorArguments']: decodeConstructorArguments(source),
-      ...parts,
-    }
-
-    return result
-  }
-}
-
-function decodeConstructorArguments(source: ContractSource): string {
-  if (source.constructorArguments.length === 0) {
-    return 'No constructor arguments'
-  }
-
-  try {
-    const abi = new utils.Interface(source.abi)
-    const fragment = abi.fragments.find((x) => x.type === 'constructor')
-    if (!fragment) {
-      return 'Constructor arguments present but no constructor found in ABI'
-    }
-
-    const decoded = toReadable(
-      utils.defaultAbiCoder.decode(
-        fragment.inputs,
-        '0x' + source.constructorArguments,
-      ),
-    )
-    const rawResult = normalizeDecodedConstructorArguments(
-      decoded,
-      fragment.inputs.length,
-    )
-
-    const structured = asStructured(rawResult, fragment.inputs)
-
-    return formatJson(structured).trim()
-  } catch {
-    return 'Error during decoding constructor arguments'
-  }
-}
-
-function normalizeDecodedConstructorArguments(
-  decoded: Readable,
-  inputLength: number,
-): Readable {
-  if (inputLength !== 1 || !Array.isArray(decoded)) {
-    return decoded
-  }
-  return decoded[0] ?? decoded
 }
 
 function splitFlatSolidity(flat: string): Record<string, string> {
@@ -217,29 +92,4 @@ function getASTTopLevelChildName(child: ASTNode): string | undefined {
       assert(false, `Unhandled child type: ${child.type}`)
     }
   }
-}
-
-export type Readable = string | number | boolean | Readable[]
-
-function toReadable(value: utils.Result): Readable {
-  if (Array.isArray(value)) {
-    return value.map(toReadable)
-  }
-  if (
-    typeof value === 'string' ||
-    typeof value === 'number' ||
-    typeof value === 'boolean'
-  ) {
-    return value
-  }
-  if (BigNumber.isBigNumber(value)) {
-    if (
-      value.gt(Number.MAX_SAFE_INTEGER.toString()) ||
-      value.lt(Number.MIN_SAFE_INTEGER.toString())
-    ) {
-      return value.toString()
-    }
-    return value.toNumber()
-  }
-  return `${value}`
 }
