+++ packages/l2b/src/implementations/discovery-ui/main.ts
@@ -13,9 +13,10 @@ import type { Server } from 'http'
 import path, { join } from 'path'
 import { attachConfigRouter } from './configs/router'
 import { DiffoveryController } from './diffovery/DiffoveryController'
+import { FlatSourceClient } from './diffovery/FlatSourceClient'
 import { attachDiffoveryRouter } from './diffovery/router'
 import { executeTerminalCommand } from './executeTerminalCommand'
-import { getCode, getCodePaths } from './getCode'
+import { getCodeFromDisk, getCodeFromEtherscan, getCodePaths } from './getCode'
 import { getConfigHealth } from './getConfigHealth'
 import { getPreview } from './getPreview'
 import { getProject } from './getProject'
@@ -85,7 +86,8 @@ export function runDiscoveryUi({ readonly }: { readonly: boolean }) {
   const templateService = new TemplateService(paths.discovery)
   const configHealthService = new ConfigHealthService()
 
-  const diffoveryController = new DiffoveryController()
+  const flatSourceClient = new FlatSourceClient()
+  const diffoveryController = new DiffoveryController(flatSourceClient)
 
   app.use(express.json())
 
@@ -133,7 +135,7 @@ export function runDiscoveryUi({ readonly }: { readonly: boolean }) {
     res.json(response)
   })
 
-  app.get('/api/projects/:project/code/:address', (req, res) => {
+  app.get('/api/projects/:project/code/:address', async (req, res) => {
     const paramsValidation = projectAddressParamsSchema.safeParse(req.params)
     if (!paramsValidation.success) {
       res.status(400).json({ errors: paramsValidation.message })
@@ -141,8 +143,16 @@ export function runDiscoveryUi({ readonly }: { readonly: boolean }) {
     }
     const { project, address } = paramsValidation.data
 
-    const checkFlatCode = readonly === false
-    const response = getCode(configReader, project, address, checkFlatCode)
+    const isLocal = readonly === false
+    const response = isLocal
+      ? getCodeFromDisk(configReader, project, address)
+      : await getCodeFromEtherscan(
+          configReader,
+          project,
+          address,
+          flatSourceClient,
+        )
+
     res.json(response)
   })
